BudgetingDesktopApplication
I am building a personal desktop application that uses MySQL as an engine for storing financial data in a database. I am studying the layouts of similar applications (that have database-like framework) and using them as guidelines. I do not intend on using every part of your program for my project. I am simply looking at it for inspiration and direction. I am new to this world and I am trying to gain better insight into the nuances of the project.